Refactor index, move db to lib
